requirements.txt
Django==5.1.9
djangorestframework==3.15.0
psycopg[binary]==3.1.17
drf-spectacular==0.26.1
python-dotenv==1.0.0
django-environ==0.11.2
django-extensions==4.0
python-dateutil==2.9.0
gunicorn==21.2.0
whitenoise==6.6.0
django-cors-headers==4.3.1
stripe==7.12.0